dependencies/alif_ml-embedded-evaluation-kit/dependencies/core-platform/requirements.txt
ethos-u-vela
pillow
pyelftools
tensorflow==2.10
